feat: move analyze_message_quality function to analysis.py and update URL routing
