+++ backend/api/urls/merge.py
@@ -3,6 +3,7 @@ from api.views.merge import (
     merge_contributions_view,
     contributors_by_category_view,
     contributors_effort_by_category_view,
+    # contributors_message_quality_view,
 )
 
 
@@ -17,5 +18,10 @@ urlpatterns = [
         "contributors_effort_by_category/",
         contributors_effort_by_category_view,
         name="contributors_effort_by_category",
+    ),
+    path(
+        "contributors_message_quality/",
+        contributors_by_category_view,  # Reusing the same view for now
+        name="contributors_message_quality",
     ),
 ]
